Liga_IA/requirements.txt
fastapi
uvicorn
numpy
scipy
pydantic
opencv-python 
mediapipe 
torch 
torchvision 
numpy 
pygame 
PyOpenGL
